requirements.txt
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
opencv-python==4.11.0.86
numpy==2.3.0
pillow==11.2.1
